types: move primary types into their own module
